float not int

+++ models/embeddings.py
@@ -115,7 +115,7 @@ class EmbeddingHelper:
     @staticmethod
     def create_embedding_matrix():
         fop = open(FLAGS.w2v_loc if FLAGS.embed_type == 0 else FLAGS.glove_loc, 'r')
-        ret = [[int(f) for f in line.split(' ')[1:]] for line in fop]
+        ret = [[float(f) for f in line.split(' ')[1:]] for line in fop]
         fop.close()
 
         return ret
